Bug 836345: Cast nsCOMArray::Count()'s result to uint32 in nsAutoCompleteController.cpp, to fix signed/unsigned comparison build warnings. r=mounir
UltraBlame original commit: 5e57a1b8f7b4f866c99af4ce4316d9e03bfc3896

+++ toolkit/components/autocomplete/nsAutoCompleteController.cpp
@@ -1071,7 +1071,9 @@ void
 nsAutoCompleteController::AfterSearches()
 {
   mResultCache.Clear();
-  if (mSearchesFailed == mSearches.Count())
+  
+  
+  if (mSearchesFailed == static_cast<uint32_t>(mSearches.Count()))
     PostSearchCleanup();
 }
 
@@ -1122,7 +1124,9 @@ nsAutoCompleteController::StartSearches()
       return rv;
     StartSearch(nsIAutoCompleteSearchDescriptor::SEARCH_TYPE_IMMEDIATE);
 
-    if (mSearches.Count() == immediateSearchesCount) {
+    
+    
+    if (static_cast<uint32_t>(mSearches.Count()) == immediateSearchesCount) {
